[REF] web: tests: extract column widths tests
